svn merge -r3747:3781 http://svn/svn/aeros/branches/granger/fix-plot-page
Summary: 

PlotPage is a real widget now, contained by PlotWindow in its tabs and the
parent of most Plot instances.  A PlotPage keeps track of the current Plot,
and a Plot draws a border highlight when its the "current" plot.
PlotWindow can install context menu actions on a Plot, eg, MultiPlotWindow
adds an action to delete the current Plot.  Lots of cleanup and
consolidation of similar menu actions among Plot classes.  The table plot
windows have been removed.

To do:

It should be straightforward now to create multiple kinds of plots within
the same MultiPlotWindow.  There just needs to be a menu for it.

Clean up the mechanisms by which PlotWindow and Plot extend each other with
menu actions.

Details (from the branch log messages):

PlotPage is now a real widget, and it's used for the tabs of the
PlotWindow; the creation of PlotWindow menus is deferred using the
aboutToShow() signal, and subclasses can override the createMenus() method
to modify the menus before display; the base createMenus() method calls
setupPlotWindow() on the current plot; the table plot windows are now
obsolete, but the Add Variable menu item is not installed into in the
PlotWindow menus

Add Variable now gets installed by the PlotQTable class on behalf of all
table plots, so this means death for the remaining table window classes

disable tabbing in CalibrationWindow

fix grid layout problem by always adding new plots either in a new row or a
new column; grid cells emptied when a plot was removed will remain empty

start on a new valgrind suppressions file for fedora7

move Plot layout management from MultiPlotWindow into PlotPage

split up creation of plot window menu in a separate virtual method, so
subclasses can add their own entries before the base class calls
Plot::setupPlotWindow()

reimplement createMenus() in MultiPlotWindow to install a menu entry for
adding a plot, which makes the same code obsolete in several Plot
subclasses

Plot carries a 'current' setting now, and PlotPage tracks the 'current'
Plot by installing an event filter and looking for clicks and keypresses in
child Plots

add PlotWindow::setupPlot() method to configure each Plot being attached
under a particular class of PlotWindow, so the PlotWindow can control the
configuration itself rather than confine it to the Plot constructor: the
PlotWindow connects signals, sets the data domain, and possibly installs
PlotActions to be used in the Plot context menus

add a PlotAction subclass of QAction which can emit a signal with the Plot
pointer as the parameter, so that Plot-specific actions from multiple Plots
can be received by the PlotWindow

eliminate PlotWindow and PlotPage from the statusMsg() chain, instead
connect the Plot's signal directly to the PlotWindow statusBar()

qwtPlotExtend no longer filters any events (ie, eventFilter() returns
false), so that PlotPage can receive the events it uses to detect changes
in the current plot

Use QFrame to control when the frame is drawn for the current plot.  I
couldn't figure out how and when to paint the frame to behave nicely with
resizes and being erased by popups and such.  So instead enable a frame,
then override drawFrame() to redraw the frame in our own special colors.  I
couldn't figure out a way to change the colors which QFrame uses, since
that requires changing the palette, and changing the palette would cause
recursive repaints.

Any PlotAction actions created as children of a Plot are installed into
popup menus by Plot::installContextMenuActions().  See
MultiPlotWindow::setupPlot(), which creates the action to delete a plot.

add delete current plot to MultiPlotWindow menu items; add handlers to
PlotPage to reset current plot when the current plot is removed or deleted