Save and restore the plot page name in configurations, and remove the plotpageindex kludge in Plot.  This change makes older configuration files incompatible.  Includes a simple test in DemoThread.
Add a flag to PlotCreateToken which specifies that a Plot should be created
empty, without any default data.  This should make it faster to restore
configurations, because the default trace does not have be loaded and then
immediately replaced with the new configuration.

Start an AddPlotDialog for adding different kinds of plots to a
MultiPlotWindow.  Not enabled in the menus yet.